server/requirements.txt
asgiref==3.11.1
attrs==26.1.0
bcrypt==5.0.0
Django==6.0.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
django-cors-headers
django-filter
djangorestframework-simplejwt
drf-spectacular
bcrypt
gunicorn==21.2.0